+++ src/Cwaff-Guns/BBGun.cs
@@ -1,34 +1,39 @@
+
 namespace CwaffingTheGungy;
 
 public class BBGun : CwaffGun
 {
     public static string ItemName         = "B. B. Gun";
     public static string ShortDescription = "Spare No One";
-    public static string LongDescription  = "Fires a single large projectile that bounces off walls and knocks enemies around with extreme force. Projectile damage and knockback scale with projectile speed. Ammo can only be regained by interacting with the projectiles once they have come to a halt.";
+    public static string LongDescription  = "Fires a single large projectile that bounces off walls and knocks enemies around with extreme force. Projectile damage and knockback scale with projectile speed. Projectiles can be caught while in motion, restoring ammo but knocking the player back in the process. Ammo can only be regained by catching projectiles, interacting with them while they are at rest, or clearing a room.";
     public static string Lore             = "This gun was originally used in the mid-18th century for hunting turkeys, as they were the only birds slow enough to actually hit with any degree of reliability. While hunters quickly decided that using a large, slow, rolling projectile wasn't ideal for hunting, the gun's legacy lives on today in shooting arenas known as \"alleys\", where sporting enthusiasts roll similar projectiles against red and white wooden objects in hopes of scoring a \"turkey\" themselves.";
 
-    private static readonly float[] _CHARGE_LEVELS  = {0.25f,0.5f,1.0f,2.0f};
-    private float                   _lastCharge     = 0.0f;
+    private const float _BOUNCE_SPEED_DECAY = 0.8f;
 
+    private static readonly float[] _CHARGE_LEVELS  = {0.25f,0.5f,1.0f,2.0f};
     internal static Projectile _PinProjectile = null;
+
+    private float _lastCharge = 0.0f;
+    private readonly List<Projectile> _extantBbs = new();
 
     public static void Init()
     {
         Lazy.SetupGun<BBGun>(ItemName, ShortDescription, LongDescription, Lore)
-          .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.CHARGE, reloadTime: 0.5f, ammo: 3, canGainAmmo: false,
+          .SetAttributes(quality: ItemQuality.B, gunClass: GunClass.CHARGE, reloadTime: 0.5f, ammo: 1, canGainAmmo: false,
             shootFps: 10, chargeFps: 16, loopChargeAt: 32, muzzleVFX: "muzzle_b_b_gun", muzzleFps: 30, muzzleScale: 0.5f, muzzleAnchor: Anchor.MiddleCenter,
             fireAudio: "Play_WPN_seriouscannon_shot_01", reloadAudio: "Play_ENM_flame_veil_01")
           .AssignGun(out Gun gun)
           .InitProjectile(GunData.New(
-            clipSize: 3, cooldown: 0.7f, angleVariance: 10.0f, shootStyle: ShootStyle.Charged, sequenceStyle: ProjectileSequenceStyle.Ordered,
+            clipSize: -1, cooldown: 0.7f, angleVariance: 10.0f, shootStyle: ShootStyle.Charged, sequenceStyle: ProjectileSequenceStyle.Ordered,
             customClip: true, speed: 20f, range: 999999f, sprite: "bball", fps: 20, anchor: Anchor.MiddleCenter, hitSound: "bb_impact_sound",
-            glowAmount: 1000f, glowColor: Color.magenta, anchorsChangeColliders: false, overrideColliderPixelSizes: new IntVector2(2, 2)))
+            glowAmount: 1000f, glowColor: Color.magenta/*, anchorsChangeColliders: false, overrideColliderPixelSizes: new IntVector2(2, 2)*/))
           .Attach<PierceProjModifier>(pierce => {
             pierce.penetration = Mathf.Max(pierce.penetration, 999);
             pierce.penetratesBreakables = true; })
           .Attach<BounceProjModifier>(bounce => {
             bounce.numberOfBounces     = Mathf.Max(bounce.numberOfBounces, 999);
             bounce.chanceToDieOnBounce = 0f;
+            bounce.percentVelocityToLoseOnBounce = 1f - _BOUNCE_SPEED_DECAY;
             bounce.onlyBounceOffTiles  = true; })
           .Attach<TheBB>()
           .CopyAllImpactVFX(Items.Crestfaller)
@@ -48,6 +53,49 @@ public class BBGun : CwaffGun
             trail.StartColor = Color.grey;
             trail.EndColor   = Color.Lerp(Color.grey, Color.white, 0.25f);
           });
+    }
+
+    public override void OnPlayerPickup(PlayerController player)
+    {
+        base.OnPlayerPickup(player);
+        player.OnRoomClearEvent += this.OnRoomClear;
+    }
+
+    public override void OnDroppedByPlayer(PlayerController player)
+    {
+        player.OnRoomClearEvent -= this.OnRoomClear;
+        base.OnDroppedByPlayer(player);
+    }
+
+    public override void OnDestroy()
+    {
+        if (this.PlayerOwner)
+            this.PlayerOwner.OnRoomClearEvent -= this.OnRoomClear;
+        base.OnDestroy();
+    }
+
+    private void OnRoomClear(PlayerController player)
+    {
+        this.gun.CurrentAmmo = this.gun.AdjustedMaxAmmo;
+        this.gun.ForceImmediateReload();
+        foreach (MiniInteractable mi in TheBB._ExtantPickups)
+            if (mi)
+                UnityEngine.Object.Destroy(mi.gameObject);
+        TheBB._ExtantPickups.Clear();
+        foreach (Projectile bb in this._extantBbs)
+            if (bb)
+            {
+                bb.gameObject.GetComponent<TheBB>()._createInteractible = false;
+                bb.DieInAir(false, false, false, true);
+            }
+        this._extantBbs.Clear();
+    }
+
+    public override void PostProcessProjectile(Projectile projectile)
+    {
+        base.PostProcessProjectile(projectile);
+        if (projectile.gameObject.GetComponent<TheBB>())
+            this._extantBbs.Add(projectile);
     }
 
     public class PinProjectile : MonoBehaviour
@@ -133,10 +181,14 @@ public class TheBB : MonoBehaviour
     private const float _BASE_EMISSION      = 3.0f;
     private const float _EXTRA_EMISSION     = 30.0f;
     private const float _BASE_ANIM_SPEED    = 2.0f;
-    private const float _BOUNCE_SPEED_DECAY = 0.9f;
     private const float _MIN_REFLECT_RADIUS = 1f;
     private const float _MAX_REFLECT_RADIUS = 2f;
     private const float _REFLECT_ANGLE_SNAP = 30f;
+    private const float _DRAG = 0.9f;
+
+    internal static readonly List<MiniInteractable> _ExtantPickups = new();
+
+    internal bool _createInteractible = true;
 
     private Projectile _projectile;
     private PlayerController _owner;
@@ -144,10 +196,12 @@ public class TheBB : MonoBehaviour
     private float _damageMult = _BB_DAMAGE_SCALE;
     private float _knockbackMult = _BB_FORCE_SCALE;
     private bool _mastered = false;
+    private bool _freebie = false;
 
     private void Start()
     {
         this._projectile = base.GetComponent<Projectile>();
+        this._projectile.baseData.damping = 0.05f;
         this._owner = this._projectile.Owner as PlayerController;
 
         if (this._owner)
@@ -155,13 +209,55 @@ public class TheBB : MonoBehaviour
             this._damageMult = _BB_DAMAGE_SCALE * this._owner.DamageMult();
             this._knockbackMult = _BB_FORCE_SCALE * this._owner.KnockbackMult();
             this._mastered = this._projectile.Mastered<BBGun>();
+            this._projectile.collidesWithPlayer = true;
+            this._projectile.UpdateCollisionMask();
+            this._projectile.specRigidbody.OnPreRigidbodyCollision += this.OnMightCollideWithPlayer;
+            _projectile.specRigidbody.m_specificCollisionExceptions.TryRemove(this._owner.specRigidbody);
         }
 
-        if (!this._projectile.FiredForFree())
-            this._projectile.OnDestruction += CreateInteractible;
+        this._freebie = this._projectile.FiredForFree();
+        if (!this._freebie)
+            this._projectile.OnDestruction += MaybeCreateInteractible;
         this._maxSpeed = this._projectile.baseData.speed;
 
         this.GetComponent<BounceProjModifier>().OnBounce += this.OnBounce;
+    }
+
+    private void OnMightCollideWithPlayer(SpeculativeRigidbody myRigidbody, PixelCollider myPixelCollider, SpeculativeRigidbody otherRigidbody, PixelCollider otherPixelCollider)
+    {
+        if (!this._owner || this._owner.specRigidbody != otherRigidbody)
+        {
+            if (otherRigidbody.gameObject.GetComponent<PlayerController>())
+                PhysicsEngine.SkipCollision = true; // don't collide with other player in coop
+            return;
+        }
+        if (!this._freebie && (this._owner.FindBaseGun<BBGun>() is Gun gun) && (gun.CurrentAmmo < gun.AdjustedMaxAmmo))
+        {
+            gun.CurrentAmmo += 1;
+            gun.ForceImmediateReload();
+        }
+        // don't do knockback if player is behind us
+        float angleToPlayer = (this._owner.CenterPosition - this._projectile.SafeCenter).ToAngle();
+        float knockbackAngle = this._projectile.specRigidbody.Velocity.ToAngle();
+        if (angleToPlayer.AbsAngleTo(knockbackAngle) < 85f)
+            otherRigidbody.gameObject.GetComponent<KnockbackDoer>().ApplyKnockback(this._projectile.specRigidbody.Velocity, this._projectile.baseData.force);
+
+        // do vfx / sfx as necessary
+        float speed = this._projectile.m_currentSpeed;
+        if (speed > 10f)
+        {
+            if (speed > 20f)
+                GameManager.Instance.MainCameraController.DoScreenShake(new ScreenShakeSettings(0.01f * speed, 6f, 0.1f, 0f), this._projectile.SafeCenter);
+            CwaffVFX.SpawnBurst(prefab: Groundhog._EarthClod, numToSpawn: 10, basePosition: this._projectile.SafeCenter, positionVariance: 1f,
+              velType: CwaffVFX.Vel.AwayRadial, velocityVariance: 0.1f * speed, lifetime: 0.5f, fadeOutTime: 0.25f,
+              startScale: 1.0f,  endScale: 0.1f, uniform: true, randomFrame: true);
+        }
+        otherRigidbody.gameObject.Play((speed > 20f) ? "bb_catch_sound" : "bb_impact_sound");
+
+        // finish up
+        this._createInteractible = false;
+        this._projectile.DieInAir();
+        PhysicsEngine.SkipCollision = true;
     }
 
     private static AIActor CheckValidEnemyOwner(Projectile proj)
@@ -217,39 +313,39 @@ public class TheBB : MonoBehaviour
 
     public void OnBounce()
     {
-        this._projectile.MultiplySpeed(_BOUNCE_SPEED_DECAY);
-        this._projectile.SendInDirection(this._projectile.m_currentDirection, resetDistance: false, updateRotation: true);
         base.gameObject.Play("bb_impact_sound");
     }
 
-    private void CreateInteractible(Projectile p)
+    private void MaybeCreateInteractible(Projectile p)
     {
+        if (!this._createInteractible)
+            return;
+
         MiniInteractable mi = MiniInteractable.CreateInteractableAtPosition(
           this._projectile.sprite,
           this._projectile.SafeCenter,
           BBInteractScript);
-            mi.doHover = true;
-            mi.sprite.SetGlowiness(glowAmount: _BASE_EMISSION, glowColor: Color.magenta);
+        mi.doHover = true;
+        mi.sprite.SetGlowiness(glowAmount: _BASE_EMISSION, glowColor: Color.magenta);
+        _ExtantPickups.Add(mi);
     }
 
     private void Update()
     {
-        float newSpeed = Mathf.Max(this._projectile.baseData.speed - _BB_SPEED_DECAY * BraveTime.DeltaTime, 0.0001f); //TODO: maybe use real friction
-
-        Material m = this._projectile.sprite.renderer.material;
-        m.SetFloat("_EmissivePower", _BASE_EMISSION + _EXTRA_EMISSION * (newSpeed / _maxSpeed));
-        m.SetFloat("_Cutoff", 0.1f);
-
-        if (newSpeed <= 1)
+        float speed = this._projectile.m_currentSpeed;
+        if (speed <= 1)
         {
             this._projectile.DieInAir(suppressInAirEffects: true);
             return;
         }
 
-        this._projectile.SetSpeed(newSpeed);
-        this._projectile.baseData.damage        = this._damageMult * newSpeed;
-        this._projectile.baseData.force         = this._knockbackMult * newSpeed;
-        this._projectile.spriteAnimator.ClipFps = Mathf.Min(_BASE_ANIM_SPEED * newSpeed, 60f);
+        Material m = this._projectile.sprite.renderer.material;
+        m.SetFloat("_EmissivePower", _BASE_EMISSION + _EXTRA_EMISSION * (speed / _maxSpeed));
+        m.SetFloat("_Cutoff", 0.1f);
+
+        this._projectile.baseData.damage        = this._damageMult * speed;
+        this._projectile.baseData.force         = this._knockbackMult * speed;
+        this._projectile.spriteAnimator.ClipFps = Mathf.Min(_BASE_ANIM_SPEED * speed, 60f);
         Lazy.PlaySoundUntilDeathOrTimeout("bb_rolling", this._projectile.gameObject, 0.1f);
 
         if (this._mastered)
@@ -263,6 +359,7 @@ public class TheBB : MonoBehaviour
             gun.CurrentAmmo += 1;
             gun.ForceImmediateReload();
             Lazy.DoPickupAt(i.sprite.WorldCenter);
+            _ExtantPickups.TryRemove(i);
             UnityEngine.Object.Destroy(i.gameObject);
         }
         i.interacting = false;